plot_network: split into separate scripts for power, hydrogen, gas
